// src/app/models/clients.model.ts
export interface Client {
    id: number;
    clientCode: string;
    firstName: string;
    lastName: string;
    dob: any;
    email: string;
    height: string;
    weight: string;
    address: string;
    city: string;
    district: string;
    town: string;
    contactFName: string;
    contactLName: string;
    gender: string;
    phone: string;
    contactPhone: string;
    password: string;
    aim: string;
    interest: string;
    healthIssues: string;
    role: string;
    image: string;
    trainer: string;
    status: string;
    file: any;
}